10477 Move Coding component reads into companion obj
