+++ test/uk/gov/hmrc/tai/model/domain/calculation/CodingComponentSpec.scala
@@ -25,8 +25,8 @@ import uk.gov.hmrc.tai.model.domain.calculation.CodingComponent.codingComponentW
 import scala.io.Source
 import scala.util.Random
 
-class CodingComponentHipReadsSpec extends PlaySpec {
-  import CodingComponentHipReadsSpec.*
+class CodingComponentSpec extends PlaySpec {
+  import CodingComponentSpec.*
 
   private val basePath = "test/resources/data/TaxAccount/CodingComponent/hip/"
   private def readFile(fileName: String): JsValue = {
@@ -42,7 +42,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
       "no NpsComponents of interest are present in the list of income deductions, within the supplied nps tax account json" in {
         val payload = readFile("tc01.json")
         payload.as[Seq[CodingComponent]](
-          CodingComponentHipReads.codingComponentReads
+          CodingComponent.codingComponentReads
         ) mustBe empty
       }
     }
@@ -52,7 +52,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
 
         val payload = readFile("tc02.json")
         payload.as[Seq[CodingComponent]](
-          CodingComponentHipReads.codingComponentReads
+          CodingComponent.codingComponentReads
         ) mustBe
           Seq(CodingComponent(PersonalAllowancePA, None, 10, "Loan Interest Amount", Some(100)))
       }
@@ -62,7 +62,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
       "single type 35 NpsComponent is present (underpayment from previous year), within income deductions" in {
         val payload = readFile("tc03.json")
         payload.as[Seq[CodingComponent]](
-          CodingComponentHipReads.codingComponentReads
+          CodingComponent.codingComponentReads
         ) mustBe
           Seq(CodingComponent(UnderPaymentFromPreviousYear, None, 10, "Van Benefit"))
       }
@@ -72,7 +72,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
       "multiple NpsComponents of interest are present within the income deductions" in {
         val payload = readFile("tc04.json")
         payload.as[Seq[CodingComponent]](
-          CodingComponentHipReads.codingComponentReads
+          CodingComponent.codingComponentReads
         ) mustBe
           Seq(
             CodingComponent(EstimatedTaxYouOweThisYear, None, 10, "Non-qualifying Relocation Expenses"),
@@ -87,7 +87,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
 
         val payload = readFile("tc05.json")
         payload.as[Seq[CodingComponent]](
-          CodingComponentHipReads.codingComponentReads
+          CodingComponent.codingComponentReads
         ) mustBe
           Seq(
             CodingComponent(EstimatedTaxYouOweThisYear, None, 10, "Non-qualifying Relocation Expenses", None),
@@ -104,7 +104,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
     "return all allowances received" when {
       "multiple allowances are present" in {
         val payload = readFile("tc06.json")
-        payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads) mustBe Seq(
+        payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads) mustBe Seq(
           CodingComponent(PersonalPensionPayments, None, 100, "Personal Pension Payments", None),
           CodingComponent(GiftAidPayments, None, 100, "Maintenance Payments", None),
           CodingComponent(EnterpriseInvestmentScheme, None, 100, "Total gift aid Payments", None),
@@ -130,7 +130,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
     "return all deductions" when {
       "multiple deductions are present" in {
         val payload = readFile("tc07.json")
-        payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads) mustBe Seq(
+        payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads) mustBe Seq(
           CodingComponent(MarriedCouplesAllowanceToWifeMAW, None, 100, "Maintenance Payments", None),
           CodingComponent(BalancingCharge, None, 100, "BPA Received from Spouse/Civil Partner", None),
           CodingComponent(UnderpaymentRestriction, None, 100, "Benefit in Kind", None),
@@ -150,7 +150,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
     "return non-tax-code incomes from income sources" when {
       "there are non-tax-code incomes present" in {
         val payload = readFile("tc08.json")
-        payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads) mustBe Seq(
+        payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads) mustBe Seq(
           CodingComponent(StatePension, None, 100, "Gift Aid Payments", None),
           CodingComponent(PublicServicesPension, None, 100, "Gift Aid treated as paid in previous tax year", None),
           CodingComponent(ForcesPension, None, 100, "One off Gift Aid Payments", None),
@@ -179,25 +179,25 @@ class CodingComponentHipReadsSpec extends PlaySpec {
     "return empty list" when {
       "there is no total liability present in tax account" in {
         val payload = readFile("tc09.json")
-        payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads) mustBe empty
+        payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads) mustBe empty
       }
       "total liability is null in tax account" in {
         val payload = readFile("tc10.json")
-        payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads) mustBe empty
+        payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads) mustBe empty
       }
       "No Benefits or Allowances were found within the provided iabd summaries" in {
         val payload = readFile("tc11.json")
-        val result = payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads)
+        val result = payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads)
         result mustBe Seq.empty[CodingComponent]
       }
       "an empty iabd summary array is present within the tax account json" in {
         val payload = readFile("tc12.json")
-        val result = payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads)
+        val result = payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads)
         result mustBe Seq.empty[CodingComponent]
       }
       "processing a solitary nps iabd json fragment that does not declare a type" in {
         val payload = readFile("tc13.json")
-        payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads) mustBe Seq
+        payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads) mustBe Seq
           .empty[CodingComponent]
       }
 
@@ -206,7 +206,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
     "return correct response" when {
       "processing test scenario 25 payload from HIP E2E testing, incl flat rate expenses and med insurance" in {
         val payload = readFile("tc34.json")
-        payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads) mustBe Seq(
+        payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads) mustBe Seq(
           CodingComponent(PersonalAllowancePA, None, 12570, "Loan Interest Amount", Some(12570)),
           CodingComponent(EarlyYearsAdjustment, None, 250, "Car Benefit", Some(250)),
           CodingComponent(StatePension, None, 10700, "Gift Aid Payments", Some(10700)),
@@ -220,7 +220,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
       "processing nps iabd summaries that correspond to known benefit types" in {
         val payload = readFile("tc14.json")
         val benefitComponents =
-          payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads)
+          payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads)
         benefitComponents.size mustBe 28
 
         benefitComponents.map(_.componentType) must contain allElementsOf Seq(
@@ -258,7 +258,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
 
         val benInKindJson = taxAccountJsonWithIabds(jsBenInKindIabdSeq)
         val benInKindComponentSeq =
-          benInKindJson.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads)
+          benInKindJson.as[Seq[CodingComponent]](CodingComponent.codingComponentReads)
         benInKindComponentSeq.size mustBe 1
         benInKindComponentSeq.head.componentType mustBe BenefitInKind
       }
@@ -266,12 +266,12 @@ class CodingComponentHipReadsSpec extends PlaySpec {
 
     "generate Benefit instances with correctly populated content" in {
       val payload1 = readFile("tc15.json")
-      payload1.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads) mustBe
+      payload1.as[Seq[CodingComponent]](CodingComponent.codingComponentReads) mustBe
         Seq(CodingComponent(CarBenefit, Some(13), 120.22, "Car Benefit"))
 
       val payload2 = readFile("tc16.json")
 
-      payload2.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads) mustBe
+      payload2.as[Seq[CodingComponent]](CodingComponent.codingComponentReads) mustBe
         Seq(CodingComponent(MedicalInsurance, None, 0, "Medical Insurance"))
     }
 
@@ -281,7 +281,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
         val payload = readFile("tc17.json")
 
         val result = payload.as[Seq[CodingComponent]](
-          CodingComponentHipReads.codingComponentReads
+          CodingComponent.codingComponentReads
         )
 
         result mustBe Seq(
@@ -292,7 +292,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
       }
       "a benefits in kind total amount is not present within iabd summaries" in {
         val payload = readFile("tc18.json")
-        val result = payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads)
+        val result = payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads)
 
         result mustBe Seq(
           CodingComponent(Accommodation, Some(1), 1, "Accommodation", None),
@@ -303,7 +303,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
       }
       "a benefits in kind total amount is present, but it has an amount value of zero" in {
         val payload = readFile("tc19.json")
-        val result = payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads)
+        val result = payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads)
 
         result mustBe Seq(
           CodingComponent(Accommodation, Some(1), 1, "Accommodation", None),
@@ -317,7 +317,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
 
       "benefits in kind total amount does not match the sum of individual benefit in kind amounts" in {
         val payload = readFile("tc20.json")
-        val result = payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads)
+        val result = payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads)
 
         result mustBe Seq(CodingComponent(BenefitInKind, Some(1), 4, "Benefit in Kind", None))
       }
@@ -326,7 +326,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
     "exclude any 'benefits from employment' from the amount reconciliation process" in {
       val payload = readFile("tc21.json")
 
-      val result = payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads)
+      val result = payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads)
 
       result mustBe Seq(
         CodingComponent(CarFuelBenefit, Some(1), 1, "Car Fuel Benefit", None),
@@ -340,7 +340,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
 
       "total and individual amounts match for all employments" in {
         val payload = readFile("tc22.json")
-        val result = payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads)
+        val result = payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads)
 
         result mustBe Seq(
           CodingComponent(Accommodation, Some(1), 1, "Accommodation", None),
@@ -366,7 +366,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
       }
       "total and individual amounts match for only one employment" in {
         val payload = readFile("tc23.json")
-        val result = payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads)
+        val result = payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads)
 
         result mustBe Seq(
           CodingComponent(Accommodation, Some(1), 1, "Accommodation", None),
@@ -378,7 +378,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
       }
       "total and individual amounts match for one employment, and total is not present for the other" in {
         val payload = readFile("tc24.json")
-        val result = payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads)
+        val result = payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads)
 
         result mustBe Seq(
           CodingComponent(Accommodation, Some(1), 1, "Accommodation", None),
@@ -404,7 +404,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
       }
       "total and individual amounts do not match for one employment, and total is not present for the other" in {
         val payload = readFile("tc25.json")
-        val result = payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads)
+        val result = payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads)
 
         result mustBe Seq(
           CodingComponent(BenefitInKind, Some(1), 30, "Benefit in Kind", None),
@@ -430,7 +430,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
 
     "Ignore a benefit in kind that does not have an employmentId" in {
       val payload = readFile("tc26.json")
-      val result = payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads)
+      val result = payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads)
 
       result mustBe Seq(
         CodingComponent(Accommodation, Some(1), 1, "Accommodation", None),
@@ -442,7 +442,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
 
     "return all benefits from employment, regardless of presence of an employmentId" in {
       val payload = readFile("tc27.json")
-      val result = payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads)
+      val result = payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads)
 
       result mustBe Seq(
         CodingComponent(ServiceBenefit, None, 1, "Service Benefit", None),
@@ -481,7 +481,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
 
       val payload = readFile("tc28.json")
       payload.as[Seq[CodingComponent]](
-        CodingComponentHipReads.codingComponentReads
+        CodingComponent.codingComponentReads
       ) must contain allElementsOf expectedCodingComponents
     }
 
@@ -490,7 +490,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
       "processing nps iabd summaries that correspond to known allowance types" in {
         val payload = readFile("tc29.json")
         val allowanceComponents =
-          payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads)
+          payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads)
 
         allowanceComponents.size mustBe 14
         allowanceComponents.map(_.componentType) mustBe
@@ -515,7 +515,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
 
     "generate Allowance instances with correctly populated content" in {
       val payload = readFile("tc30.json")
-      payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads) mustBe
+      payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads) mustBe
         Seq(CodingComponent(JobExpenses, Some(13), 120.22, "Job Expenses", None))
 
       taxAccountJsonWithIabds(
@@ -525,7 +525,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
             "estimatesPaySource" -> 1
           )
         )
-      ).as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads) mustBe
+      ).as[Seq[CodingComponent]](CodingComponent.codingComponentReads) mustBe
         Seq(CodingComponent(HotelAndMealExpenses, None, 0, "test", None))
 
     }
@@ -540,14 +540,14 @@ class CodingComponentHipReadsSpec extends PlaySpec {
 
       val payload = readFile("tc31.json")
       payload.as[Seq[CodingComponent]](
-        CodingComponentHipReads.codingComponentReads
+        CodingComponent.codingComponentReads
       ) must contain allElementsOf expectedCodingComponents
     }
 
     "codingComponentReads" must {
       "merge the coding components coming from incomeSourceReads and totalLiabilityReads into a new list" in {
         val payload = readFile("tc32.json")
-        payload.as[Seq[CodingComponent]](CodingComponentHipReads.codingComponentReads) mustBe
+        payload.as[Seq[CodingComponent]](CodingComponent.codingComponentReads) mustBe
           Seq(
             CodingComponent(PersonalAllowancePA, None, 10, "Loan Interest Amount", None),
             CodingComponent(CommunityInvestmentTaxCredit, Some(1), 1, "Community Investment Tax Credit", None),
@@ -584,7 +584,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
 
         val payload = readFile("tc33.json")
         payload.as[Seq[CodingComponent]](
-          CodingComponentHipReads.codingComponentReads
+          CodingComponent.codingComponentReads
         ) must contain allElementsOf expectedCodingComponents
       }
 
@@ -666,7 +666,7 @@ class CodingComponentHipReadsSpec extends PlaySpec {
   }
 }
 
-object CodingComponentHipReadsSpec {
+object CodingComponentSpec {
   private val nino: Nino = new Generator(new Random).nextNino
 
   private def npsIabdSummary(iadbType: Int): JsObject =
